# Conflicts:
#	src/haven/GameUI.java
#	src/haven/Gob.java
#	src/haven/LocalMiniMap.java
#	src/haven/MapView.java
#	src/haven/MenuGrid.java
#	src/purus/DragonflyCollector.java
#	src/purus/MusselPicker.java

+++ src/haven/Message.java
@@ -269,9 +269,9 @@ public abstract class Message {
                 case T_NIL:
                     ret.add(null);
                     break;
-	    case T_UID:
-		ret.add(int64());
-		break;
+                case T_UID:
+                    ret.add(int64());
+                    break;
                 case T_BYTES:
                     int len = uint8();
                     if ((len & 128) != 0)
@@ -338,11 +338,12 @@ public abstract class Message {
     public Message adduint32(long num) {
         int off = wget(4);
         Utils.uint32e(num, wbuf, off);
-	return(this);
+        return (this);
     }
+
     public Message addint64(long num) {
-	int off = wget(8);
-	Utils.int64e(num, wbuf, off);
+        int off = wget(8);
+        Utils.int64e(num, wbuf, off);
         return (this);
     }
 
@@ -403,7 +404,7 @@ public abstract class Message {
                 addfloat32(((Float) o).floatValue());
             } else if (o instanceof Double) {
                 adduint8(T_FLOAT64);
-                addfloat64(((Float) o).floatValue());
+                addfloat64(((Double) o).floatValue());
             } else {
                 throw (new RuntimeException("Cannot encode a " + o.getClass() + " as TTO"));
             }
